Crate descartes

Structs

Area

AreaSplitResult

Band

BoundaryPiece

BoundingBox

ClosedLinePath

ConcatError

CurvedPath

DashIterator

Intersection

LinePath

LineSegment

PrimitiveArea
Represents a filled area bounded by a clockwise boundary. Everything "right of" the boundary is considered "inside"

ShiftVector

UnclosedPathError

Enums

AreaError

AreaLocation

CurvedSegment

PieceEquivalence

PieceRole

Constants

THICKNESS

Traits

HasBoundingBox

Intersect

Into2d

Into3d

PointContainer

RoughEq

WithUniqueOrthogonal

Functions

angle_along_to

angle_to

signed_angle_to

try_inverse
Tries to gets an inverted copy of a square matrix.

Type Definitions

Aff3

Iso3

M4

N

P2

P3

Persp3

V2

V3

V4

VecLike